+++ demo/scripts/main.all.js
@@ -1,2 +1,2 @@
-/* updated : 2013-05-31 */
+/* updated : 2013-06-16 */
 myapp.set("foo","foofoo"),myapp.set({bar:"barbar"}),myapp.set("baz",function(){return"bazbaz"}),function(){var a=document.getElementById("log");a.innerHTML=[myapp.get("foo"),myapp.get("bar"),myapp.get("baz")()].join("\n")}();