+++ src/head-require.js
@@ -93,7 +93,7 @@
 	};
 
 	global[loader.namespace] = app;
-	global.headRequire = function(){
+	global.headRequire = head.require = function(){
 		loader.require.apply(loader, arguments);
 	};